feat: 옷 등록 과정 간소화에 의한 옷 수정 수정 (#519)
